requirements.txt
requests~=2.31.0
cryptography
PyYAML~=6.0
pycryptodome
prettytable
